+++ src/index.js
@@ -15,6 +15,8 @@ export { ViewportScrollScene } from './ViewportScrollScene'
 export { ViewportScrollScene as PerspectiveCameraScene } from './ViewportScrollScene'
 export { useDelayedCanvas } from './useDelayedCanvas'
 
+export { HijackedScrollbar } from './scrollbar/HijackedScrollbar'
+
 // Public & battle-tested
 // ----------------------------------
 export { VirtualScrollbar } from './scrollbar/VirtualScrollbar'